Update client.lua
In GetBrokenATM i have moved the return statement outside of the if statement so that it returns false as default if the decoration does not exist on the entity then moved the RequestNamedPtfxAsset function call outside of the loop to prevent unnecessary repetition of the request. combined street variable into a single line, the if statement to check the atm and its health into a single statement, all to reduce the amount of checks performed and simplify the code.